veyron2/vom2: Fix any(?foo) encoding, and add a test.
The vom2 encoder had a bug where it encoded any(?foo)
incorrectly.  The correct encoding has the order:
  type ID     [for concrete type of value stored in any]
  exists byte [for the optional foo]

The buggy encoding flipped the order around:
  exists byte [for the optional foo]
  type ID     [for concrete type of value stored in any]

This is obvious after you know the bug, but note that it was a
bit tricky to track down.  The exists byte is set to 1 to mean
"exists", which also happens to be the type ID for the "any"
type.  So the buggy encoding would almost be decoded correctly,
and would only fail on some values of the subsequent data.

Added a test for this case, which exposed the problem and the
fix.  The original problem was discovered by running the store
tests in vom2 mode.

Change-Id: Ia6a515f03ed5faa8dec786a0cebc5f63997dd7ff