veyron2/vdl/valconv: Fix bug wrt unexported struct fields.
Suharsh reported a bug he noticed in vom2 here:
  https://github.com/veyron/release-issues/issues/608

The bug is in valconv FromReflect.  If a Go struct has unexported
fields, those fields don't appear in the VDL type; that is
correct.  But FromReflect needs to account for this possible
mismatch when we're converting from a Go struct.

Also added some tests to both valconv and vom2 to expose these
kinds of issues.

Change-Id: Ic5c059660042090db4e8073c69655af7dce78165