veyron2/vdl: Add support for Go "native types" in VDL.
Here's the high-level idea.  VDL/VOM provide mechanisms to ensure
wire compatibility across different programming environments.
The VDL type system is fairly complex, to allow users to model
their APIs in a natural fashion.

But sometimes it's not enough; the user defines a type in VDL to
ensure wire compatibility, but really wants to use a different
type in their native generated API that's more convenient.  A
simple example is the Go standard time.Time and time.Duration.
The user wants to uses these standard types in their API, but
still wants wire compatibility.

That's the problem that Go "native types" is solving.  The idea
is that the user can add a stanza to the vdl.config file in their
vdl package, which tells the vdl compiler about the native type.
Then they add a pair of conversion functions converting between
the vdl wire type and the Go native type.  Thereafter the VDL
value conversion and vom will know about the native type, and
will automatically marshal the native type into the wire type
when serializing, and also unmarshal in the other direction.

Details:

1) The conversion functions are defined as methods on the wire
type.  In theory we could have defined the methods on the native
type instead, but that wouldn't work for Go standard time.Time.
This is also a bit nicer, since VDL codegen automatically adds
type registration in init funcs for all types defined in VDL
files, and the same registration mechanism is used.

2) If the wire type isn't registered, we'll never know about the
native type, and the automatic marshaling won't happen.  This
means we probably want to use this mechanism sparingly.

3) The vdl compiler changes haven't been implemented yet; this CL
contains all the value conversion / reflection logic.  I'll make
the vdl compiler changes in a separate CL.

4) This CL also includes minor changes to the vdl parser, to make
it emit better error messages for config files.  Every config
file for the vdl tool is named "vdl.config", so it's nice to also
output the package path in the error message, in order to figure
out where particular failures are happening.

5) Also removed the Go NoFmt option in vdl.config; we always
gofmt generated Go source files.

Change-Id: I9d6a4c2f5a4de58f4dc3778cfd7849bf35a707ff