veyron2/vdl: Add library funcs for compiling/parsing expressions.
The main library function is BuildExprs in the vdl/build package.
This combines parse.ParseExprs and compile.CompileExpr, taking a
string with one or more VDL expressions, and turning them into a
slice of *vdl.Value.  This will be used in an upcoming change to
allow vrpc to take VDL expressions as input.

There isn't support for specifying types from other packages yet;
that'll be added as a change to the VDL syntax to allow
fully-specified package paths.

Change-Id: I22ce0e31db9d0623b2529be18961838a4a1c3d39