veyron2/vdl: Allow package-path qualified identifiers in VDL.
Before this change, imports were either referred to by their
package name, or the local identifier chosen by the user:

  package foo

  import (
    "a/b/bar"
    zzz "c/d/baz"
  )

  type A1 bar.T
  type B2 zzz.T

After this change, we also allow the full-package path to be used
to identify an import:

  package foo

  import (
    "a/b/bar"
    zzz "c/d/baz"
  )

  type A1 bar.T
  type A2 "a/b/bar".T // same as bar.T

  type B1 zzz.T
  type B2 "c/d/baz".T // same as zzz.T

This is useful when parsing vdl values specified outside of VDL
files.  E.g. the vrpc tool will soon use VDL syntax to specify
arbitrary values, and it's useful to have non-ambiguous syntax to
refer to other packages.

NOTE: now that we have a way to disambiguate imports that have
conflicting package names, we can consider removing support for
user-chosen local import identifiers.  This support has been left
intact for now, since the general logic is the same as for
regular package qualified identifiers.

Change-Id: I8590fea478fc215f3e82e08dfde82d6cb2c95988