veyron2/ipc,veyron/runtimes/google/ipc: remove ipc.LeafDispatcher
An attempt at deprecating ipc.LeafDispatcher. I don't feel very strongly about
this change, but all things being equal, I prefer to keep the public API surface
as small as possible. With the advent of the server.Serve API, there are few
reasons left for developers to need something like LeafDispatcher. In fact, the
only legitimate use cases left were in unit tests (where it's sometimes
convenient to define helper routines that take in a dispatcher), and in the
server.Serve implementation.

This CL moves LeafDispatcher to the testutil package so it can be used in
tests. server.Serve implements leaf dispatcher internally, and also extends the
Serve API to also accept an invoker (which is unrelated, but seems consistent
with how Dispatcher.Lookup is allowed to behave).

Change-Id: Ic129374a4e8a869854e4c944ce962ed84b70450f
MultiPart: 1/4