veyron/tools/vrpc: Update to new __Signature and VDL values.
The original vrpc tool was still using the old Signature support,
and specified values using the old vom/JSON support, both of
which are deprecated and need to be removed.

The new vrpc implemented in this CL uses the new __Signature
support, and specifies values using VDL syntax.

There are remaining TODOs to clean up the formatting of the
output, and handle streaming input arguments.  Those will come in
separate follow-up CLs; this CL was getting too big already, and
is necessary as a first-step so that we can remove the deprecated
__Signature and vom1 support.

Change-Id: Ia65d5be4a046dd8bec381bfbcbd9a02472ac7199
MultiPart: 1/3