core: Remove vom1, wiretype and old Signature.
This includes vdl tool changes to stop generating the old
Signature methods, as well as vdl file updates.

There are non-trivial changes in:
  veyron/runtimes/google/ipc
    Remove old vom1 logic and unused vom2 transition code.
  veyron/lib/exec
    Replace vom with vom2, fix pointer receiver bug so that the
    same RWMutex is used.
  veyron/services/mgmt/device/impl
    Change proxyInvoker to use Describe__ rather than Signature.

veyron/lib/exec: Change config to use vom2 for serialization.

Also fixed a bug where the cfg wasn't being passed as a pointer
method receiver.  The problem with the old code is that the
RWMutex would be different on each method invocation, making it
useless.

Change-Id: I208751b52adfaeaa1bdd95207c1f39d3ff03bc60
MultiPart: 1/6