core: Replace vdlutil.Register with vdl.Register.
The vdlutil.Register function was only meant for the transition
from vom1 to vom2; now that the transition is over, we can just
call vdl.Register instead.

Also note that types defined in vdl files are automatically
registered in a package init function, so some calls can be
removed completely.

Change-Id: I7812a47427cee02d232c668ed7cb24f9d1981573
MultiPart: 1/2