core: Remove 'errorid' from all *.vdl files.
This is step 1 in a multi-step change to add VDL support for
verror2, which is itself a step in the transition from verror1 to
verror2.

Note that 'errorid' is our VDL support for verror1, which will be
replaced by new 'error' syntax for verror2.

Here are the anticipated steps in the transition:
  1) Remove 'errorid' from VDL files.
  2) Remove 'errorid' support from vdl tool.
  3) Add new 'error' support to vdl parser/compiler.
  4a) Add Go codegen support for new 'error' support.
  4b) Add Javascript codegen support for new 'error' support.
  4c) Add Java codegen support for new 'error' support.
  5) Use new 'error' support in VDL files.

The strategy for this CL is simple: remove 'errorid' from *.vdl
files, and add manual code that creates the verror.ID instead.

This means that Go code ends up working the same as it was
before.  Javascript doesn't generate code for 'errorid' yet, so
also works the same as it was before.  Java may break, but we can
fix that up separately, since it's not on the 0.1 release path.

Change-Id: Id6f4d16f7f7fcb92b1ac1281e3737e2c6c23f063