veyron/services/mgmt/device/impl: add Debug RPC method for sundry state dumps
The way we currently debug the device manager is by poking around the directory
tree that the device manager sets up on the local filesystem. Sometimes, this is
not convenient (e.g. when we're remote), and other times the information we want
is not available on the filesystem (e.g. the credentials for app instances, when
using the security agent).

This CL adds the new Debug method to the Application interface in the VDL.

The corresponding implementation supports app installations and instances for
now, and prints an assortment of stuff about the respective installations and
instances.

Change-Id: I84dd88ff0324f103590651fdfca3252f17d37da7