"core": Make go-tests pass under agent
This CL ensures that all tests under v.io/core/... pass when
run under an agent and do not pollute the agent principal's state.

In particular,
- It adds a functon 'InitForTest' to the testutil package that invokes
veyron2.Init() and returns the initialized context after setting a freshly
created principal on it. The principal setting step is skipped if the
function is invoked from a process created by the modules framework.
- It replaces all usages of veyron2.Init() in tests with
  testutil.InitForTest()
Change-Id: Ia5cc1cd5dba7e443da686591a42e22da45b9d98a