vdl/javascript: Generate vdl errors in the javascript code.
This change also includes a new flag based on discussions with Nick and
Benj about how to deal with cyclic dependencies of generated code.  We
have the issue that JS core code depends on VDL generated files and VDL
generated files depend on JS core code.  We decided that there will be
two ways that we generated imports for JS generated files.  The default
mode assumes that the files will be generated outside of the JS core
repo and will require('veyron').  This will work for all cases that
aren't in js/core.  For js/core we'll use relative paths to import the
portions of js/core that the generated files need.  This allows us to
merge the JS core repo and the JS VOM repo.  Right now because there is a
large change outstanding on the VOM repo, we can't merge the two, so
all the support isn't implemented, but will be by the time we release.

MultiPart: 1/2
Change-Id: Ib4f2b8d21d4d8bcda328985df36451ec0041a628