veyron2/vdl: Change Register to be recursive.
Also make TypeFromReflect call Register automatically.  This is
needed to avoid some of the subtle ordering issues that crop up,
since Register, TypeOf and TypeFromReflect are often called in
either init functions or in variable initializers.

Some more details here:
https://github.com/veyron/release-issues/issues/907

The changes in valconv.Convert and convert_test.go are bugfixes, which were
exposed since the tests now register types and we expect to return real go
objects.  Previously most of the types weren't registered, and we'd return
*vdl.Value instead.

Change-Id: Ia4c6e672c6b330a6b71aa1fc07074a7913208d72