vdl/javascript: Append 'Error' to the end of VDL errors.
The new convention for errors is to not have Error in the id because
different languages have different conventions for how to name errors.
For an error named Foo:
  go -> ErrFoo
  js -> FooError
  java -> FooException
To make this possible and nice, we have the code generations attach the
right strings to make it natural in the target language.

MultiPart: 1/2
Change-Id: Ib88af2028ca4f4ffc757e7a0fc12060e1dca7106