veyron2/vdl: Remove valconv dependency on verror package.
This is step N-2 of a multi-step process to refactor the
vdl/... packages.  The problem with the vdl/... packages is that
functions that are seemingly related are not grouped together in
the same packages.  E.g. we have vdl.TypeOf, but valconv.ValueOf.

This step removes the valconv dependency on verror, by
introducing yet-another-registration-mechanism.  In theory we
might have been able to refactor the wire<->native type
registration mechanism in vdl.Register to achieve this, but it
gets complicated since go error is an interface, and thus needs
some special-cases when performing reflection.

At this point both vdl and vdl/valconv have no dependencies on
other packages under v.io.  The valconv tests still have a
dependency on verror; we'll move the tests into the valconv_test
package in a separate CL.

The next step will be to refactor the vdl and vdl/valconv
packages, either by combining into a single vdl package, or
moving functions around between the packages.

Change-Id: I40026249b75cf25e4da66524f1154ae055c5fec7