TBR: verror: add BadEtag error (migrated from access package)
Main motivation: The group server uses etags to cover both ACLs and
group data. I.e. the scope of an etag is broader than just the ACL.
(And, we anticipate using this pattern for other servers in the
future.)

Secondary motivation: In the future we may want to use etags for other
things entirely.

Aside: I think we should consider renaming "etag" to something else,
but I'd prefer to save that debate for later. :)

Change-Id: I160e518c93655e84c554f63b57073d613df05865