v.io/core/veyron2/verror: Don't export verror's Params() and SubErrors() functions.
[ Shyam, Srdjan:  You might want to do something similar in Java/Javascript. ]

Dave observed that the ability for programmers easily to access
the parameter list and subordinate error list in a verror error
is an attractive nuisance:  people will use them as general
purpose mechanisms for returning things, when they should
instead use other mechanisms.

This CL discourages the abuse by making the Go functions
verror.Params() and verror.SubErrors() no longer be exported
(i.e., be error.params() and verror.subErrors()).  It adds
comments to make people think before they undo this change.  The
calls are not used outside of the verror package, so this should
have no impact on client code.

The calls are used internally in verror, so I opted to make them
private rather removing them and moving the code.

We don't expect to need to reverse this change, but it's easy to
do so should we find that we were hasty.  In contrast, it would
be hard to make this change later if we left the functions as
public.

Beware that programmers still could access the raw ParamList
field of the Standard type; the field is exported so that it will be
sent over the wire.  This may change if the wire type acquires a
special encoding.

Tested:
        cd release/go/src/v.io
        v23 go test ./...
        v23 test run vanadium-go-test
The only errors already exist, and are in the old store code,
due to it no longer being maintained.  After today it's expected that
its tests will not be included in the standard tests.

Change-Id: I4b283d7a00f09cbf954fff975a85136a3b2c2ab0