TBR veyron/runtimes/google/ipc/stream/fix: fix a bug in vif cache
Currently, vifs are cached with the actual protocol not with a
client specified protocol. E.g., vif with "wsh" are cached with
"tcp" or "ws". So we cannot find a vif with "wsh" causing unncessary
vif creation. There is also a similar problem with v4 and v6 protocols.
This CL addresses this issue.

MultiPart: 2/2
Change-Id: I04052a6b8f44da422ba92f877d301ace47d2ec31