v23/vom/testdata: Add more conversion tests
This adds more cross-language conversion tests to vomdata.vdl.config

While doing this, I have come to the conclusion that the current setup
is not expressive enough.

However, the current tests are enough to show that the JS-side's
conversion function works the way we want it to.

I plan to refactor by:
https://github.com/veyron/release-issues/issues/1188
- splitting off vomdata.vdl.config into separate files
- modifying generate.go to generate different data structures
  depending on the vomFOO.vdl.config file received
- change the organization of the conversion tests
- possibly change the organization of the compatibility tests

MultiPart: 2/2
Change-Id: I83fb9aebef9bfd509884c1d5670a2bf6580c2ae4