vdl: Invoke RegisterNative once per type (instead of once per .vdl file).
Prior to this change, the RegisterNative call would be placed in the
init function for each generated .vdl.go file. Thus, if there were
multiple .vdl files in a package, there would be a runtime panic
about RegisterNative being called twice.

With this change, the RegisterNative call is made only in the file
that defines the wire type.

Unfortunately, I couldn't figure out a good way to test this out,
but I observed this when trying to use this native conversion
in the v23/security package (which has multiple .vdl files).

Change-Id: I59db26be902474944f7b82fbf1bcd504cce70ca5