v23: Change vdl.WireError format; flatten IDAction and RetryCode.
This is a backwards-incompatible change to the standard VDL error
wire format.  The Go verror.Standard in-memory representation is
left unchanged; m3b will update that in a future CL.

The main change is to remove the nested IDAction struct.  That
was legacy behavior from a convenience in the original Go verror2
implementation, which shouldn't be propagated to the wire format
for all languages.

The Action is also changed from a bit-packed uint32 to a
WireRetryCode enum.  The enum is nicer as a wire format, since it
is convertible to a string, and provides better debuggability in
Javascript.  As a reminder, enums are still encoded with an
integer index in the binary vom format; they just take more space
on the wire encoding.

I also discovered that our WireError<->error conversions were
previously depending on the fact that verror.Standard and
vdl.WireError were identical for all exported fields; I'd missed
some conversion special-cases.  The tests showed me the problem,
and I've fixed that up.

MultiPart: 1/2
Change-Id: I1e35bfd3290053216d87353bb7334c24408faafd