TBR v23: Re-org vdl and vom packages.
The main idea is that everything under v.io/v23/vdl/... has moved
to v.io/core/veyron/..., except for the top-level v.io/v23/vdl
package, which stays in v23.  The sub-packages consist of the vdl
tool, and libraries to help implement the tool.

The vom and vomtestgen tools also moved to v.io/core/veyron/...

This makes the dependencies worse (temporarily), because of the
vdlroot/signature package, which pulls in a bunch of
dependencies.  That will be re-org'd separately.

After that step, we'll be left with
"github.com/cosmosnicolaou/llog" as our only external dep.

MultiPart: 1/7
Change-Id: I5bf0d7d1c0ee3ef50e7bd145a7e06966134a59dd