v.io/v23/verror: Change types to match wire format.
This CL and two companion CLs change verror types to more
closely resemble their VDL wire formats.

Javascript, the fact that IDAction was a struct nested inside an error
  was an inconvenience.  So that's being flattened into two fields within the struct.

Additionally, verror.Standard is renamed to verror.E

[We had thought of changing ID to Id, which we had to do in the wire format;
without that, the Javascript name becomes "iD" which looks hideous.  But
changing this in the Go source conflicts with the Go team's mysterious desire
to spell "Id" as "ID", even though it's not an acronym.  So we're leaving "ID"
as it is in Go.]

See
	https://vanadium-review.googlesource.com/#/c/6034/
for the corresponding changes in core

MultiPart: 1/2
Change-Id: I885656eeff62e81454209e41427c29819f299c3c