naming: Endpoint version 4.
Robin suggested keeping the server's blessings in the endpoint and
it seems that doing so will simplify a lot of the server authorization
setup. Mount and Resolve etc. will no longer have to explicitly keep
track of server blessings but instead it will naturally follow from
the server's address (endpoint).

This is step 1 of many in an attempt to make that happen.
Doing so required reserving the "," from blessing names, but that
should be fine (and we should be reserving more substrings for
future use such as for groupservers etc.)

MultiPart: 1/2
Change-Id: I64fbd2fed2fdd3b5edc5da14cc36c5d398f75a44