v23/vdl: Fix native type conversion bugs, and add time.Deadline
One bug in vdl.Convert was simple; we were missing a case to
convert from wire types back to native types when the conversion
target is an interface{}.  That caused decoding of native types
to always return their zero value.

The other bug was complex; we weren't able to use the native type
support, where the wire type is a vdl.Union.  This is tricky
because vdl.Union is itself generated as an interface, so there
are many cases to handle.

Here's the issue tracking the above bugs:
https://github.com/veyron/release-issues/issues/1194

In addition, added standard VDL type time.WireDeadline, with an
associated native type time.Deadline.  The concept is that the
native type uses absolute time to represent deadlines, while the
wire type uses a duration from "now".  This is more robust in the
face of time skew, and will be used in at least two places: the
ipc.Request header for the request timeout, and Dave's MountEntry
for TTLs.
https://github.com/veyron/release-issues/issues/1073

It may seem like these are somewhat unrelated changes; they are
related in that the native type conversion bugs must be fixed in
order for any native types to be useful, including deadlines.
I've mixed them into 1 CL because I have less than 20 hours
before I leave, so I'd like to get this stuff submitted.

Change-Id: I7d511c02db294bf18be853c2a6b54c81d6ba2e11