v.io/v23/config: Convert uses of fmt.Errorf to verror.
This change converts all uses of fmt.Errorf in v.io/v23/config
to use verror.  The purpose is to make the code base more homogenous
so that later developers are more likely to follow suit, and to make it possible
to render the error messages in other languages, should this be necessary.

Error message strings are preserved by registering a local verror error
for each error format string.  (Presise details may have changed, such as server names
being prefixed to the error messages.)
I fixed a test that was pattern matching on error messages by instead checking for error codes.

Aside:  I often could have used a context.T when calling verror.New().
It would be good if more of these calls took context.T in the public APIs.

Change-Id: Ic827cb50c6fa6aefa1be9b37408919c69565a3da