v.io/v23/security: Convert uses of fmt.Errorf to verror.
This change converts all uses of fmt.Errorf in v.io/v23/security
to use verror.  The purpose is to make the code base more homogenous
so that later developers are more likely to follow suit, and to make it possible
to render the error messages in other languages, should this be necessary.

Error message strings are preserved by registering a local verror error
for each error format string.  (Presise details may have changed, such as server names
being prefixed to the error messages, or it formatting other than default was requested.
I fixed these only when the change was egregious, or a test failed.)

Aside:  I often could have used a context.T when calling verror.New().
It would be good if more of these calls took context.T in the public APIs.

See also:
	https://vanadium-review.googlesource.com/6962
which fixes v.io/v.io/x/ref/security/audit/principal_test.go to
accommodate changes in the implementation of the errors it now receives;
it was using reflect.DeepEqual on them.

MultiPart: 1/2
Change-Id: I3f297b1e6957563a107d82a071e8c2cfa92ea3c9