services/mgmt/device: Change application.Start()
With this change, the Start() method allows the caller to bless the
application instance directly instead of blessing the device manager,
which then had to issue another blessing for the app instance.

The new interface streams the instances name and public key to the
client, and the client streams the blessings back to the server.

An added benefit is that the instance names can be sent to the client as
soon as they are created, even if the Start call ultimately fails.

MultiPart: 1/2
Change-Id: I7b8e3bf79108fa8533dbf1dec4248f48e222f28b