TBR: security: Bugfix.
(A presubmit test on the non-TBRed change passed, so I decided
to TBR to unblock some other submits)

If Blessings.chains had a capacity larger than length, then one call to
Bless might mess up previous calls to it as it modifies the same
underlying array element (Certificate) as used in a previous call.

I observed this bug in another test, but am having trouble creating a
unittest that causes the underlying chain to have a capacity larger than
length.

Change-Id: Id8fda84a873980e195b4d03d9d56a5d73ff5a2cf