vtrace: Vtrace should never cause panics.
Several people have complained about needing to initialize vtrace
in tests to prevent a panic when the user doesn't care about vtrace
output.  Vtrace is a debugging tool, so it doesn't really make sense
that it should ever be the cause of a panic.

Instead we generate empty vtrace types that silently do nothing.

Change-Id: I55a591ac08559b4fe8b9fbd9d8a02caf724b2cab