v23: Disallow uppercase acronyms in VDL identifiers.
The purpose of this change is to ensure code in each generated
language is idiomatic and looks reasonable.  The problem with
uppercase acronyms is they either look bad or are non-reversible,
when converted to camelCase or CONST_CASE.

E.g. EOF is converted to eOF, which looks bad.  MyTTL is
converted to MY_TTL, which is non-reversible.

Const identifiers are currently special-cased to avoid this rule.
In the future we may choose to change to the same rule for const
identifiers; this is less important since const identifiers are
never sent over the wire.

MultiPart: 1/3
Change-Id: Ibcb4631e918a0f63e111a8e621034133bbeed0a1