"v23/security": BlessingNames -> [Local/Remote]BlessingNames
CL 7951 introduces the security.BlessingNames API with an explicit
'side' argument. This API is somewhat ugly as it requires referring
to security.CallSide[Local/Remote] at all call sites. Furthermore,
we realize that we are never interested in validating caveats on the
local blessings, and therefore, we would never want to invoke
BlessingNames with CallSideLocal. Instead we would just invoke BlessingsInfo
on call.LocalBlessings.

In light of the above findings, this CL deprecates BlesssingNames and
instead adds the following functions
- func RemoteBlessingNames(call) // equivalent to BlessingNames(call,
CallSideRemote)
- func LocalBlessingNames(call) // equivalent to
  call.LocalPrincipal().BlessingsInfo(call.LocalBlessings())

MultiPart: 1/2

Change-Id: I61ea2426902a557ee90677c36f0e17ef1cde4c2f