"v23/security": Remove security.CallSide
We recently added the security.CallSide paramater to CaveatValidators
in order to support validating caveats on both local and remote
blessings. However, we soon realized that we really don't have a usecase for
validating caveats on local blessings. It suffices to simply use
principal.BlessingInfo to determine the set of names
encapsulated within local blessings.

Therefore in order to keep the APIs simple, this CL gets rid of
security.CallSide. Caveat validation is always carried out assuming that
the credential came from the remote end.

MultiPart: 1/3
Change-Id: I06c95f69444d8e5f1d0d19d919a8258d016eeb1d