services/security/access: PermissionsAuthorizer will not work with multiple access tags defined on the method.
Some folks feel it will be more natural to require the set of blessings
to have permissions to *all* access tags on a method.

Others feel that it will be more natural to require the set of blessings
to have permissions on *any* access tag on a method.

Yet others feel that having multiple tags on a method is just too
confusing.

The cynical argue that if there are multiple tags, why not support
arbitrary boolean expressions involving them.

Bogdan suggested the conservative first approach: Disallow multiple tags
and then choose the appropriate handling when we have more mileage on
the system and there is a clearer winner of which approach to take.

Closes veyron/release-issues#1500

Change-Id: I83a5669aec25f770921c8397f665eb6fd6f7291b