vio/v23/vom: Use verror instead of fmt.Errors and errors.New
This change uses verror instead of fmt.Errors and errors.New
in the v23/vom code.

Format strings were changed as little as possible---typically
just adding {1:}{2:} at the start.
One exception was
	errUnexpectedControlByte
in vom/decoder.go which had the format 0x%x and now uses {3},
which means the byte will be in decimal.   We could force to hex
by using sprintf("0x%x", ...) as the parameter, if desired.

Change-Id: I84a4f148c2b4cc1bcbe6b6dd63dccd1a2cbc311f